# Remove Allocations from Path.Encode