Removed a bunch of the allocations for big speed improvements:

```
benchmark               old ns/op     new ns/op     delta
BenchmarkPathEncode     135389        23444         -82.68%
```

@mlerner